removed references to helloworld, refactored to brewpub
