+++ src/main/java/org/launchcode/brewpub/models/Review.java
@@ -1,8 +1,7 @@
-package org.launchcode.helloworld.models;
+package org.launchcode.brewpub.models;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.ManyToOne;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
