+++ src/main/java/org/launchcode/brewpub/models/data/ReviewRepository.java
@@ -1,6 +1,6 @@
-package org.launchcode.helloworld.models.data;
+package org.launchcode.brewpub.models.data;
 
-import org.launchcode.helloworld.models.Review;
+import org.launchcode.brewpub.models.Review;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
