+++ src/main/resources/templates/fragments.html
@@ -6,7 +6,7 @@
 
     <link th:href="@{/css/bootstrap.css}" rel="stylesheet" />
     <link th:href="@{/css/brewpub.css}" rel="stylesheet" />
-    <script type="text/javascript" th:src="@{/js/bootstrap.js}"></script>
+<!--    <script type="text/javascript" th:src="@{/js/bootstrap.js}"></script>-->
 
 </head>
