+++ src/test/java/org/launchcode/brewpub/BrewPubApplicationTests.java
@@ -1,4 +1,4 @@
-package org.launchcode.helloworld;
+package org.launchcode.brewpub;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
